Move new dependent functional tests out of user folder
