+++ Testing/functional/tests/cypress/integration/e2e/dependent/management.js
@@ -1,8 +1,8 @@
-const { AuthMethod } = require("../../../../support/constants");
+const { AuthMethod } = require("../../../support/constants");
 import {
     getCardSelector,
     getTabButtonSelector,
-} from "../../../../support/functions/dependent";
+} from "../../../support/functions/dependent";
 
 const validDependent = {
     firstName: "Sam ", // Append space to ensure field is trimmed
